backend/requirements.txt
fastapi[standard]
pydantic-extra-types==2.11.1
pydantic-settings==2.13.1